revised demo and retrieve_rerank method
