+++ rank_llm/run_rank_llm.py
@@ -1,6 +1,4 @@
 import argparse
-import json
-from pathlib import Path
 import sys
 import os
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -9,20 +7,10 @@ sys.path.append(os.path.dirname(SCRIPT_DIR))
 import torch
 
 from rank_llm.pyserini_retriever import RetrievalMethod
-from rank_llm.rank_gpt import SafeOpenai
 from rank_llm.rankllm import PromptMode
-from rank_llm.rank_vicuna import RankVicuna
 from rank_llm.topics_dict import TOPICS
-from rank_llm.trec_eval import EvalFunction
-from rank_llm.retrieve_and_rerank import RetrievalMode, Retriever, Reranker
-
-
-def get_api_key() -> str:
-    from dotenv import dotenv_values, load_dotenv
-    import os
-
-    load_dotenv(dotenv_path=f"../.env.local")
-    return os.getenv("OPEN_AI_API_KEY")
+from rank_llm.retriever import RetrievalMode
+from rank_llm.retrieve_and_rerank import retrieve_and_rerank
 
 
 def main(args):
@@ -36,64 +24,21 @@ def main(args):
     shuffle_candidates = args.shuffle_candidates
     print_prompts_responses = args.print_prompts_responses
     device = "cuda" if torch.cuda.is_available() else "cpu"
-    if "gpt" in model_path:
-        openai_keys = get_api_key()
-        agent = SafeOpenai(
-            model=model_path,
-            context_size=context_size,
-            prompt_mode=prompt_mode,
-            keys=openai_keys,
-        )
-    else:
-        agent = RankVicuna(
-            model=model_path,
-            context_size=context_size,
-            prompt_mode=prompt_mode,
-            device=device,
-            num_gpus=num_gpus,
-        )
-    candidates_file = Path(
-        f"../retrieve_results/{retrieval_method.name}/retrieve_results_{dataset}.json"
-    )
-    if not candidates_file.is_file():
-        print("Retrieving:")
-        retriever = Retriever(RetrievalMode.DATASET)
-        retriever.retrieve(dataset=dataset, retrieval_method=retrieval_method)
-    else:
-        print("Reusing existing retrieved results.")
+    retrieval_mode = RetrievalMode.DATASET
 
-    with open(candidates_file, "r") as f:
-        retrieved_results = json.load(f)
-
-    print("Reranking:")
-    reranker = Reranker(agent, top_k_candidates, dataset)
-    (
-        rerank_results,
-        input_token_counts,
-        output_token_counts,
-        aggregated_prompts,
-        aggregated_responses,
-    ) = reranker.rerank(
-        retrieved_results, 
-        rank_end=top_k_candidates, 
-        window_size=20,
-        shuffle_candidates=shuffle_candidates,
-        logging=print_prompts_responses)
-
-    file_name = reranker.write_rerank_results(
-        retrieval_method.name,
-        rerank_results,
-        input_token_counts,
-        output_token_counts,
-        aggregated_prompts,
-        aggregated_responses,
+    rerank_results = retrieve_and_rerank(
+        model_path,
+        top_k_candidates,
+        dataset,
+        retrieval_mode,
+        retrieval_method,
+        context_size,
+        device,
+        num_gpus,
+        prompt_mode,
         shuffle_candidates,
+        print_prompts_responses
     )
-
-    print("Evaluating:")
-    EvalFunction.eval(["-c", "-m", "ndcg_cut.1", TOPICS[dataset], file_name])
-    EvalFunction.eval(["-c", "-m", "ndcg_cut.5", TOPICS[dataset], file_name])
-    EvalFunction.eval(["-c", "-m", "ndcg_cut.10", TOPICS[dataset], file_name])
 
 
 """ sample run:
